Split hybrid retrieval helpers
