+++ cognee/modules/retrieval/hybrid/ranking.py
@@ -1,0 +1,47 @@
+from cognee.modules.retrieval.hybrid.results import payload, result_id
+
+
+def rank_chunk_summary_pairs(
+    pairs: list[dict],
+    limit: int,
+    use_importance_weight: bool,
+) -> list[dict]:
+    if limit <= 0:
+        return []
+
+    rrf_k = _rrf_k(limit)
+    ranked = []
+    for pair in pairs:
+        chunk = pair["chunk"]
+        if chunk is None:
+            continue
+
+        ranks = [
+            rank
+            for rank in (pair["bm25_rank"], pair["vector_rank"], pair["summary_rank"])
+            if rank is not None
+        ]
+        if not ranks:
+            continue
+
+        rrf_score = sum(1.0 / (rrf_k + rank + 1) for rank in ranks)
+        final_score = rrf_score
+        if use_importance_weight:
+            final_score *= _importance_factor(chunk)
+
+        chunk_id = pair["chunk_id"] or result_id(chunk) or ""
+        ranked.append((final_score, rrf_score, min(ranks), chunk_id, pair))
+
+    ranked.sort(key=lambda item: (-item[0], -item[1], item[2], item[3]))
+    return [pair for *_, pair in ranked[:limit]]
+
+
+def _rrf_k(chunks_top_k: int) -> int:
+    return max(30, min(60, 20 + 2 * chunks_top_k))
+
+
+def _importance_factor(chunk) -> float:
+    raw_importance = payload(chunk).get("importance_weight")
+    importance = raw_importance if isinstance(raw_importance, (int, float)) else 0.5
+    importance = max(0.0, min(1.0, importance))
+    return 0.75 + 0.5 * importance
